// Desktop/cs5500-gui-arch-huangweili97-main/src/types/types.tsx
export type QuestionsFilterType = { qSlice: Question[], qLength: number }
// ...
export interface AnswerParamType
 {
    aid: string;
    text: string;
    ansBy: string;
    ansDate: Date;
}
// ...
export interface QuestionParamType {
    qid: string;
    title: string;
    text: string;
    tagIds: string[];
    askedBy: string;
    askDate: Date;
    ansIds: string[];
    views: number;
}

/**
 * An interface representing the input parameters
 * the constructor of the Tag class expects
 */
export interface TagParamType {
    tid: string;
    name: string;
}

/**
 * An interface representing the input parameters
 * the constructor of the Application class expects in the model
 */
export interface ApplicationProps {
    questions: QuestionParamType[];
    tags: TagParamType[];
    answers: AnswerParamType[];
}

/**
 * An interface for the model's Application class
 */
export interface ApplicationInterface {
    setPageInstance(arg0: any): unknown;
    questions: QuestionParamType[];
    tags: TagParamType[];
    answers: AnswerParamType[];
    addAnswer: (qid: string, answer: { text: string; ansBy: string }) => string;
    addQuestion: (question: {
        title: string;
        text: string;
        askedBy: string;
        tags: string[];
    }) => string;
    addTag: (tagname: string) => string;
    getQuestionCountByTag: (tid: string) => number;
    getQuestionsByFilter: (startIndex: number, order?: string, search?: string) => QuestionsFilterType;
    getQuestionById: (qid: string | undefined) => Question | undefined;
    getQuestionAnswer: (question: Question | null) => Answer[];
    getTagCount: () => number;
    getTags: () => Tag[];
    getTagById: (id: string) => Tag | null;
}